Modified extract_twop_baryons_h5.sh to support the High Momenta Form

+++ scripts/twop/extract_twop_baryons_h5.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-if [ $# -ne 10 ]
+if [ $# -ne 11 ]
 then
-    echo "Usage: $0 <extract_dir> <h5_file-prefix> <out_dir> <src_list> <mom_list> <conf> <T> <Qsq> <exe_dir>  <baryon, see list below>"
+    echo "Usage: $0 <extract_dir> <h5_file-prefix> <out_dir> <src_list> <mom_list> <conf> <T> <Qsq> <exe_dir> <HighMomForm:1 LowMomForm:0> <baryon, see list below>"
     printf "  0: all\n  1: nucl_nucl\n  2: nucl_roper\n  3: roper_nucl\n  4: roper_roper\n  5: deltapp_deltamm_11\n  6: deltapp_deltamm_22\n  7: deltapp_deltamm_33\n  8: deltap_deltaz_11\n  9: deltap_deltaz_22\n 10: deltap_deltaz_33\n"
     exit
 fi
@@ -16,7 +16,8 @@ CONF=$6
 T=$7
 Qsq=$8
 EXE_DIR=$9
-bar=${10}
+HighMomForm=${10}
+bar=${11}
 
 LIST[1]="nucl_nucl"
 LIST[2]="nucl_roper"
@@ -39,29 +40,67 @@ then
     BARYON_LIST=${LIST[$bar]}
 fi
 
-while read x y z t
-do 
-    FILE=${DIR}/${H5_PRE}.${CONF}_baryons_Qsq${Qsq}_SS.${x}.${y}.${z}.${t}.h5
 
-    src=sx${x}sy${y}sz${z}st${t}
-    EX_DIR=${DIR}/${src}
-    mkdir -p ${EX_DIR}
+if [ $HighMomForm -eq 0 ]
+then
 
-    OUT_FILE=${OUT_DIR}/twop.${CONF}.baryons.SS.${x}.${y}.${z}.${t}.dat
+    while read x y z t
+    do 
+	FILE=${DIR}/${H5_PRE}.${CONF}_baryons_Qsq${Qsq}_SS.${x}.${y}.${z}.${t}.h5
 
-    rm -f ${OUT_FILE}
+	src=sx${x}sy${y}sz${z}st${t}
+	EX_DIR=${DIR}/${src}
+	mkdir -p ${EX_DIR}
 
-    for tp in ${BARYON_LIST}
-    do
-	while read mx my mz
+	OUT_FILE=${OUT_DIR}/twop.${CONF}.baryons.SS.${x}.${y}.${z}.${t}.dat
+
+	rm -f ${OUT_FILE}
+
+	for tp in ${BARYON_LIST}
 	do
-	    EX_FILE=${EX_DIR}/twop.${CONF}.baryons.${tp}.${mx}_${my}_${mz}.SS.${x}.${y}.${z}.${t}.dat
-	    
-	    ${EXE_DIR}/extract_twop_baryons_h5 ${FILE} ${EX_FILE} ${tp} ${mx} ${my} ${mz} ${CONF} ${src} ${T} 0
+	    while read mx my mz
+	    do
+		EX_FILE=${EX_DIR}/twop.${CONF}.baryons.${tp}.${mx}_${my}_${mz}.SS.${x}.${y}.${z}.${t}.dat
+		
+		${EXE_DIR}/extract_twop_baryons_h5 ${FILE} ${EX_FILE} ${tp} ${mx} ${my} ${mz} ${CONF} ${src} ${T} 0
+
+		cat ${EX_FILE} >> ${OUT_FILE}
+
+	    done < ${MOM_LIST}
+	done
+
+	rm -rf ${EX_DIR}
+
+    done < ${SRC_LIST}
+
+fi
+
+if [ $HighMomForm -eq 1 ]
+then
+
+    while read x y z t
+    do 
+	FILE=${DIR}/${H5_PRE}.${CONF}_baryons_Qsq${Qsq}_SS.${x}.${y}.${z}.${t}.h5
+
+	src=sx${x}sy${y}sz${z}st${t}
+	EX_DIR=${DIR}/${src}
+	mkdir -p ${EX_DIR}
+
+	OUT_FILE=${OUT_DIR}/twop.${CONF}.baryons.SS.${x}.${y}.${z}.${t}.dat
+
+	rm -f ${OUT_FILE}
+
+	for tp in ${BARYON_LIST}
+	do
+	    EX_FILE=${EX_DIR}/twop.${CONF}.baryons.${tp}.SS.${x}.${y}.${z}.${t}.dat
+		
+	    ${EXE_DIR}/extract_twop_baryons_HighMomForm_h5 ${FILE} ${EX_FILE} ${tp} ${CONF} ${src} ${T} 0
 
 	    cat ${EX_FILE} >> ${OUT_FILE}
+	done
 
-	done < ${MOM_LIST}
-    done
+	rm -rf ${EX_DIR}
 
-done < ${SRC_LIST}
+    done < ${SRC_LIST}
+
+fi
